Some work on msvc64 compat. Not complete

+++ codemp/win32/win_wndproc.cpp
@@ -489,7 +489,7 @@ LONG WINAPI MainWndProc (
 	case WM_INPUT:
 		{
 			RAWINPUT raw;
-			size_t rawSize = sizeof(raw);
+			UINT rawSize = sizeof(raw);
 
 			GetRawInputData( (HRAWINPUT) lParam, RID_INPUT, &raw, &rawSize, sizeof(RAWINPUTHEADER) );
